Refactor flow queries into mixin to meet LOC limit
